update darwin? and windows? broke out and made a run for 'OS' country
