+++ lib/boom.rb
@@ -10,6 +10,7 @@ require 'json/pure'
 
 $:.unshift File.join(File.dirname(__FILE__), *%w[.. lib])
 
+require 'boom/os'
 require 'boom/color'
 require 'boom/platform'
 require 'boom/command'
