+++ test/test_os.rb
@@ -1,0 +1,15 @@
+require 'helper'
+
+class TestPlatform < Test::Unit::TestCase
+
+  def setup
+  end
+
+  def test_darwin
+    assert_equal Boom::OS.darwin?, RUBY_PLATFORM.include?('darwin')
+  end
+
+  def test_windows
+    assert_equal Boom::OS.windows?, true if RUBY_PLATFORM =~ /win|mingw/ 
+  end
+end
